Move auth helper fn into utils/auth
